[Refactor] UserService 분리, interface 구조사용 - #97
