+++ src/main/java/me/snaptime/user/service/impl/SignServiceImpl.java
@@ -1,0 +1,83 @@
+package me.snaptime.user.service.impl;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import me.snaptime.exception.CustomException;
+import me.snaptime.exception.ExceptionCode;
+import me.snaptime.jwt.JwtProvider;
+import me.snaptime.profilePhoto.domain.ProfilePhoto;
+import me.snaptime.profilePhoto.repository.ProfilePhotoRepository;
+import me.snaptime.user.domain.User;
+import me.snaptime.user.dto.req.SignInReqDto;
+import me.snaptime.user.dto.req.UserReqDto;
+import me.snaptime.user.dto.res.SignInResDto;
+import me.snaptime.user.dto.res.UserResDto;
+import me.snaptime.user.repository.UserRepository;
+import me.snaptime.user.service.SignService;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Collections;
+
+@Slf4j
+@RequiredArgsConstructor
+@Service
+@Transactional
+public class SignServiceImpl implements SignService {
+
+    private final UserRepository userRepository;
+    private final ProfilePhotoRepository profilePhotoRepository;
+    private final PasswordEncoder passwordEncoder;
+    private final JwtProvider jwtProvider;
+
+    @Override
+    public UserResDto signUp(UserReqDto userRequestDto) {
+
+        //로그인 id가 이미 존재하는지 확인
+        if(userRepository.findByLoginId(userRequestDto.loginId()).isPresent()){
+            throw new CustomException(ExceptionCode.LOGIN_ID_ALREADY_EXIST);
+        }
+
+        String fileName = "default.png";
+        String filePath =  "/test_resource/" + fileName;
+
+        ProfilePhoto profilePhoto = ProfilePhoto.builder()
+                .profilePhotoName(fileName)
+                .profilePhotoPath(filePath)
+                .build();
+        //기본 프로필 저장
+        profilePhotoRepository.save(profilePhoto);
+
+        //새로운 사용자 객체 생성
+        User user = User.builder()
+                .name(userRequestDto.name())
+                .loginId(userRequestDto.loginId())
+                .password(passwordEncoder.encode(userRequestDto.password()))
+                .email(userRequestDto.email())
+                .birthDay(userRequestDto.birthDay())
+                //단일 권한을 가진 리스트 생성, 하나의 요소를 가진 불변의 리스트 생성
+                .roles(Collections.singletonList("ROLE_USER"))
+                .profilePhoto(profilePhoto)
+                .build();
+
+        return UserResDto.toDto(userRepository.save(user));
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public SignInResDto signIn(SignInReqDto signInRequestDto) {
+        User user = userRepository.findByLoginId(signInRequestDto.loginId()).orElseThrow(() -> new CustomException(ExceptionCode.USER_NOT_EXIST));
+
+        if (!passwordEncoder.matches(signInRequestDto.password(), user.getPassword())) {
+            throw new CustomException(ExceptionCode.PASSWORD_NOT_EQUAL);
+        }
+        String accessToken = jwtProvider.createAccessToken(user.getLoginId(), user.getRoles());
+
+        SignInResDto signInResponseDto = SignInResDto.builder()
+                .accessToken(accessToken)
+                .build();
+
+        return signInResponseDto;
+    }
+}
